[XrdCl] Prevent concurrent destruction and initialization of PostMaster
At the end of a client process, we stop the PostMaster to join worker
threads to avoid a situation where worker threads try to use OpenSSL
while it's already being destroyed. However, in several places in the
code, the PostMaster is initialized on demand, such as in socket
callbacks, etc, by calling XrdCl::DefaultEnv::GetPostMaster(). This
initialization atomically gets the PostMaster if initialized, or takes
a lock and tries to initialize it otherwise. When we destroy the
PostMaster, we must also take this lock to prevent such on demand
initialization to happen while a PostMaster is being destroyed.