[CMake] Place build artifacts in ${CMAKE_BINARY_DIR}/{bin,lib}
This makes the placement of binaries and libraries independent of
the directory structure of the CMake build system, which makes it
easier to set paths to client and server binaries in tests, as well
as ensuring that libraries and plugins from the build itself are
always used when running the test suite. Moreover, this allows to
migrate the build system to use more idiomatic CMake by using, for
example, add_subdirectory(Xrd*) and src/Xrd*/CMakeLists.txt instead
of the current include(Xrd*) and src/Xrd*.cmake, to keep plugins in
the same place as other libraries.