[XrdCl] Use kXR_mkpath instead of kXR_mkdir in do_Open()
This is the same kind of bug as has been fixed by commit b84d81a.
The kXR_mkdir enum is for a command, the flag we want here is
actually kXR_mkpath. However, as noted in the other commit as well,
what this means is that historically the client has always had this
effectively enabled, so as in b84d81a, we accept also kXR_async in
here, to keep backward compatibility.

Fixes: #2393, 76108afd218350d655fd0b87b4394b759abe5357